feat: Finish tests for users route

+++ backend/app/crud/users.py
@@ -37,10 +37,33 @@ def create_user(user: user.UserCreate, db: Session):
 
 def get_user(id: int, user_id: int, db: Session):
     user = db.query(User).filter(User.id == id, User.id == user_id).first()
+    
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User not found")
+
     return user
 
 def update_user(id: int, user: user.UserCreate, user_id: int, db: Session):
+    existing_user_username = db.query(User).filter(User.username == user.username, User.id != user_id).first()
+    if existing_user_username:
+        raise HTTPException(
+            status_code=status.HTTP_409_CONFLICT,
+            detail="Username is already taken",
+        )
+    
+    existing_user_email  = db.query(User).filter(User.email == user.email, User.id != user_id).first()
+    if existing_user_email:
+        raise HTTPException(
+            status_code=status.HTTP_409_CONFLICT,
+            detail="Email is already registered",
+        )
+    
     user_query = db.query(User).filter(User.id == id, User.id == user_id)
+
+    if not user_query.first():
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail="User not found")
+    
     user_query.update(user.model_dump(exclude={"email", "password"}), synchronize_session=False)
     db.commit()
     updated_user = user_query.first()
@@ -48,5 +71,10 @@ def update_user(id: int, user: user.UserCreate, user_id: int, db: Session):
 
 def delete_user(id: int, user_id: int, db: Session):
     user_query = db.query(User).filter(User.id == id, User.id == user_id)
+
+    if not user_query.first():
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail="User not found")
+
     user_query.delete(synchronize_session=False)
     db.commit()
